added interaction to user page

+++ files_blueprint.py
@@ -15,14 +15,11 @@ files_blueprint = Blueprint('example_blueprint', __name__)
 @files_blueprint.route('/files/<file_index>/<file_name>')
 def index(file_index, file_name):
     file_name = file_name.rsplit('-', 1)[0]
-    print(file_name)
     file = app.mongo.db[FILES_COLLECTION].find_one({"_id": ObjectId(file_index), "file_name": file_name})
-    print(file)
     if file is None:
         return abort(404)
     if not file["is_active"]:
         return abort(404)
-    print(file)
     return render_template("display_file.html", file=file, hash=file_index)
 
 
@@ -32,6 +29,10 @@ def download(file_index, file_name):
         return redirect("/login")
     file = app.mongo.db[FILES_COLLECTION].find_one({"_id": ObjectId(file_index),
                                                     "file_name": file_name.rsplit('.', 1)[0]})
+    if file is None:
+        return abort(404)
+    if not file["is_active"]:
+        return abort(404)
     return send_file(file["file_path"], attachment_filename=file_name)
 
 
@@ -40,11 +41,15 @@ def delete_file(file_index, file_name):
     if not usrm.check_user(request, app.mongo.db[AUTH_COLLECTION]):
         return redirect("/login")
     file_name = file_name.rsplit('-', 1)[0]
-    print(file_name)
     file = app.mongo.db[FILES_COLLECTION].find_one({"_id": ObjectId(file_index), "file_name": file_name})
+    if file is None:
+        return abort(404)
+    if not file["is_active"]:
+        return abort(404)
     if request.cookies.get("email") != file["email"]:
         return abort(403)
     os.remove(file["file_path"])
+    app.mongo.db[FILES_COLLECTION].update_one({"_id": file["_id"]}, {"$set": {"is_active": False}})
     return "delete successful"
 
 
@@ -58,7 +63,7 @@ def upload_file():
         file_name = secure_filename(f.filename)
         file_type = app.fm.get_file_type(file_name)
         file_name = file_name.rsplit('.', 1)[0]
-        file_size = request.form["size"]
+        file_size = int(request.form["size"])
         size_consumed = user.get("size_consumed", 0)
         if size_consumed + file_size > MAX_FILE_SIZE:
             return render_template(
@@ -72,6 +77,8 @@ def upload_file():
         if not os.path.exists(f"files/{user['email']}"):
             os.mkdir(f"files/{user['email']}")
         f.save(file_path)
+        # if size send by browser and original file size differ something went wrong
+        # (unsuccessful upload or user manipulate data)
         if file_size != os.stat(file_path).st_size:
             os.remove(file_path)
             return abort(400)
@@ -89,3 +96,4 @@ def upload_file():
                                                            "updated_at": datetime.now()}})
         return redirect("/dashboard")
     return render_template("files_upload.html")
+
